Issue #52 Moved the memoizing functions out of the public API module to a package protected, it's fairly tuned to a very specific use-case
